move `spliceChangesIntoString` into `utils/` folder
